1/telegram_bot/requirements.txt
mysql-connector-python==8.0.21
numpy==1.18.0
pyTelegramBotAPI==3.7.2

